Remove DoNotLoad usage, tweak fallback logic
If we touch the cassette and F5, debug, etc. or the flag gets unset somehow, the player is screwed. So instead of using Session.DoNotLoad, just remove the cassette if the target flag has been set. That way the cutscene can be triggered again if something went wrong

Also syncs flag collect to end of ground pound